+++ go.mod
@@ -1,3 +1,3 @@
-module GoDB
+module github.com/MattLaidlaw/GoDB
 
 go 1.16
